requirements.txt
colorama
python-docx
python-pptx
openpyxl
